@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:infinite_horizons/domain/study_type_abstract.dart';
 import 'package:infinite_horizons/presentation/atoms/atoms.dart';
+import 'package:infinite_horizons/presentation/molecules/molecules.dart';
 
 class EnergySelectionMolecule extends StatefulWidget {
   const EnergySelectionMolecule(this.callback);
@@ -44,25 +45,33 @@ class _EnergySelectionMoleculeState extends State<EnergySelectionMolecule> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const TextAtom('classic_pomodoro'),
-        const SeparatorAtom(),
-        energyWidget(EnergyType.medium),
-        const SeparatorAtom(variant: SeparatorVariant.farApart),
-        const TextAtom('custom'),
-        const SeparatorAtom(),
-        Column(
-          children: [
-            energyWidget(EnergyType.max),
-            energyWidget(EnergyType.veryHigh),
-            energyWidget(EnergyType.high),
-            energyWidget(EnergyType.low),
-            energyWidget(EnergyType.veryLow),
-          ],
-        ),
-      ],
+    return MarginedExpandedAtom(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const TopBarMolecule(
+            title: 'energy',
+            topBarType: TopBarType.none,
+            margin: false,
+          ),
+          const SeparatorAtom(variant: SeparatorVariant.farApart),
+          const TextAtom('classic_pomodoro'),
+          const SeparatorAtom(),
+          energyWidget(EnergyType.medium),
+          const SeparatorAtom(variant: SeparatorVariant.farApart),
+          const TextAtom('custom'),
+          const SeparatorAtom(),
+          Column(
+            children: [
+              energyWidget(EnergyType.max),
+              energyWidget(EnergyType.veryHigh),
+              energyWidget(EnergyType.high),
+              energyWidget(EnergyType.low),
+              energyWidget(EnergyType.veryLow),
+            ],
+          ),
+        ],
+      ),
     );
   }
 }
