@@ -1,6 +1,9 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:infinite_horizons/domain/study_type_abstract.dart';
 import 'package:infinite_horizons/presentation/atoms/atoms.dart';
+import 'package:infinite_horizons/presentation/molecules/molecules.dart';
 
 class MotivationOrganism extends StatelessWidget {
   const MotivationOrganism(this.callback);
@@ -28,17 +31,34 @@ class MotivationOrganism extends StatelessWidget {
       case EnergyType.max:
         text = "So much energy 🔋⚡🔋⚡🔋⚡🔋\nLet's begin";
     }
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        TextAtom(text),
-        const SeparatorAtom(variant: SeparatorVariant.farApart),
-        ButtonAtom(
-          variant: ButtonVariant.primary,
-          onPressed: callback,
-          text: 'start',
+
+    return Expanded(
+      child: MarginedExpandedAtom(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const TopBarMolecule(
+              title: 'lets_start',
+              topBarType: TopBarType.none,
+              margin: false,
+            ),
+            Expanded(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  TextAtom(text),
+                  const SeparatorAtom(variant: SeparatorVariant.farApart),
+                  ButtonAtom(
+                    variant: ButtonVariant.primary,
+                    onPressed: callback,
+                    text: 'start',
+                  ),
+                ],
+              ),
+            ),
+          ],
         ),
-      ],
+      ),
     );
   }
 }
